NutriSeekerapril copy 2/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
requests==2.31.0
pillow==10.4.0
gradio==4.7.1
pandas==2.0.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
matplotlib==3.7.2
numpy==1.24.3
transformers==4.35.0
torch==2.0.1
pyngrok==7.0.0
python-multipart==0.0.6